requirements.txt
boto3
pandas
PyYAML
requests
